// src/app/interfaces/persona.ts
export interface Persona {
	idTipoDocu: string;
	nroDocu: string;
	nombres: string;
	apePater: string;
	apeMater: string;
	fecNacimiento: string;
	strFecNacimiento: string;
	idSexo: string;
	imagen: string;
	direccion: string;
	idEstadoCivil: string;
	correoElect: string;
	telefono: string;
	estado: string;
}